removed depreciated files and merged to runmanagercore
